adding to balances store
