+++ src/routes/+layout.svelte
@@ -7,6 +7,12 @@
 	import { browser } from '$app/environment';
 	import { PUBLIC_LAUNCHED } from '$env/static/public';
 	import { flare } from '@wagmi/core/chains';
+	import Footer from '$lib/components/Footer.svelte';
+	import { ZeroAddress } from 'ethers';
+	import { cysFlrAddress } from '$lib/stores';
+	import { simulateErc20PriceOracleReceiptVaultPreviewDeposit } from '../generated';
+	import balancesStore from '$lib/balancesStore';
+	import { onDestroy, onMount } from 'svelte';
 
 	const initWallet = async () => {
 		const erckit = defaultConfig({
@@ -22,6 +28,34 @@
 	$: if (browser && window.navigator) {
 		initWallet();
 	}
+
+	let intervalId: ReturnType<typeof setInterval>;
+
+	onMount(() => {
+		startGettingPriceRatio();
+	});
+
+	const getPriceRatio = async () => {
+		const { result } = await simulateErc20PriceOracleReceiptVaultPreviewDeposit($wagmiConfig, {
+			address: $cysFlrAddress,
+			args: [BigInt(1e18), 0n],
+			account: ZeroAddress as `0x${string}`
+		});
+		balancesStore.updateLockPrice(result);
+	};
+
+	const startGettingPriceRatio = async () => {
+		intervalId = setInterval(getPriceRatio, 5000);
+		getPriceRatio();
+	};
+
+	function stopGettingPriceRatio() {
+		clearInterval(intervalId);
+	}
+
+	onDestroy(() => {
+		stopGettingPriceRatio();
+	});
 </script>
 
 {#if $wagmiConfig}
@@ -30,6 +64,6 @@
 		<main class="flex-grow bg-[#1C02B8]">
 			<slot></slot>
 		</main>
-		<!-- <Footer /> -->
+		<Footer />
 	</div>
 {/if}
